Mac: explain how to avoid an IP address clash
The issue [docker/for-mac#1803] describes a situation where the internal
IP addresses used by Docker for Mac clashed with IP addresses in use on
the local environment. This problem became worse in 17.06 because we
expanded the number of IP addresses we reserved, making clashes more likely.

This patch describes how to work around a clash by changing the internal
IP addresses used by Docker for Mac.

Signed-off-by: David Scott <dave.scott@docker.com>